slideshow/pyproject.toml
[project]
name = "slideshow"
version = "0.3.1"
description = "A feature-rich image slideshow application with Tkinter."
readme = "README.md"
requires-python = ">=3.9,<=3.11"
dependencies = [
    "pillow",
    "coloredlogs",
]
authors = [
    { name = "Your Name" },
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["slideshow"]

[project.scripts]
slideshow = "slideshow.__main__:main"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
    "pytest-cov",
]
